[refactor] datastore 모듈 api-impl 분리 (#312)
interface 와 impl 이 한 모듈내에 있는 것을 분리
