+++ app/build.gradle.kts
@@ -70,7 +70,8 @@ dependencies {
         projects.core.data.api,
         projects.core.data.impl,
         projects.core.database,
-        projects.core.datastore,
+        projects.core.datastore.api,
+        projects.core.datastore.impl,
         projects.core.designsystem,
         projects.core.model,
         projects.core.navigation,
