+++ core/datastore/api/build.gradle.kts
@@ -1,0 +1,15 @@
+plugins {
+    alias(libs.plugins.unifest.android.library)
+}
+
+android {
+    namespace = "com.unifest.android.core.datastore.api"
+}
+
+dependencies {
+    implementations(
+        projects.core.model,
+
+        libs.kotlinx.coroutines.core,
+    )
+}
