+++ core/datastore/api/src/main/kotlin/com/unifest/android/core/datastore/api/RecentLikedFestivalDataSource.kt
@@ -1,4 +1,4 @@
-package com.unifest.android.core.datastore
+package com.unifest.android.core.datastore.api
 
 import com.unifest.android.core.model.FestivalModel
 import kotlinx.coroutines.flow.Flow
