+++ core/datastore/api/src/main/kotlin/com/unifest/android/core/datastore/api/model/SettingsData.kt
@@ -1,4 +1,4 @@
-package com.unifest.android.core.datastore.model
+package com.unifest.android.core.datastore.api.model
 
 data class SettingsData(
     val isClusteringEnabled: Boolean,
