+++ settings.gradle.kts
@@ -30,7 +30,8 @@ include(
     ":core:data:api",
     ":core:data:impl",
     ":core:database",
-    ":core:datastore",
+    ":core:datastore:api",
+    ":core:datastore:impl",
     ":core:designsystem",
     ":core:model",
     ":core:navigation",
